backend/requirements.txt
APScheduler==3.6.3
asciimatics==1.14.0
asgiref==3.6.0
atomicwrites==1.4.1
attrs==22.1.0
Brotli==1.0.9
cachetools==4.2.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.1
cssselect2==0.7.0
defusedxml==0.7.1
Django==3.2.3
django-cors-headers==3.13.0
django-extra-fields==3.0.2
django-filter==22.1
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
exceptiongroup==1.0.0
flake8==6.0.0
flake8-isort==6.0.0
fonttools==4.40.0
future==0.18.2
html5lib==1.1
idna==3.4
importlite==0.1.0
iniconfig==1.1.1
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
mccabe==0.7.0
mypy==0.982
mypy-extensions==0.4.3
oauthlib==3.2.2
packaging==21.3
Pillow==9.3.0
pluggy==0.13.1
psycopg2-binary==2.9.3
py==1.11.0
pycodestyle==2.10.0
pycparser==2.21
pydyf==0.7.0
pyfiglet==0.8.post1
pyflakes==3.0.1
PyJWT==2.1.0
pyparsing==3.0.9
pyphen==0.14.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dotenv==1.0.0
python-telegram-bot==13.7
python3-openid==3.2.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.2
sqlparse==0.4.3
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tornado==6.2
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.14
wcwidth==0.2.5
weasyprint==59.0
webencodings==0.5.1
zopfli==0.2.2
django_environ==0.10.0